feat: package into executable jar
