+++ pom.xml
@@ -15,9 +15,12 @@
         <!-- Dependencies Versions -->
         <org.openjfx-version>21.0.2</org.openjfx-version>
         <org.junit.jupiter-version>5.10.2</org.junit.jupiter-version>
-        <maven-surefire-plugin-version>3.2.3</maven-surefire-plugin-version>
-        <maven-compiler-plugin-version>3.10.1</maven-compiler-plugin-version>
+
+        <!-- Plugins Versions -->
+        <maven-shade-plugin-version>3.5.2</maven-shade-plugin-version>
         <javafx-maven-plugin-version>0.0.8</javafx-maven-plugin-version>
+        <maven-surefire-plugin-version>3.2.5</maven-surefire-plugin-version>
+        <maven-compiler-plugin-version>3.12.1</maven-compiler-plugin-version>
         <maven-checkstyle-plugin-version>3.3.1</maven-checkstyle-plugin-version>
     </properties>
 
@@ -75,12 +78,12 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-surefire-plugin</artifactId>
-                    <version>3.2.5</version>
+                    <version>${maven-surefire-plugin-version}</version>
                 </plugin>
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.12.1</version>
+                    <version>${maven-compiler-plugin-version}</version>
                     <configuration>
                         <source>11</source>
                         <target>11</target>
@@ -89,7 +92,7 @@
                 <plugin>
                     <groupId>org.openjfx</groupId>
                     <artifactId>javafx-maven-plugin</artifactId>
-                    <version>0.0.8</version>
+                    <version>${javafx-maven-plugin-version}</version>
                     <executions>
                         <execution>
                             <!-- Default configuration for running with: mvn clean javafx:run -->
@@ -127,6 +130,30 @@
                         </execution>
                     </executions>
                 </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-shade-plugin</artifactId>
+                    <version>${maven-shade-plugin-version}</version>
+                    <executions>
+                        <execution>
+                            <phase>package</phase>
+                            <goals>
+                                <goal>shade</goal>
+                            </goals>
+                        </execution>
+                    </executions>
+                    <configuration>
+                        <createDependencyReducedPom>false</createDependencyReducedPom>
+                        <transformers>
+                            <transformer
+                                    implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                                <manifestEntries>
+                                    <Main-Class>osmosis.folder.inspector.Main</Main-Class>
+                                </manifestEntries>
+                            </transformer>
+                        </transformers>
+                    </configuration>
+                </plugin>
             </plugins>
         </pluginManagement>
         <plugins>
@@ -145,6 +172,10 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
             </plugin>
         </plugins>
     </build>
