+++ src/main/java/osmosis/folder/inspector/Main.java
@@ -1,25 +1,10 @@
 package osmosis.folder.inspector;
 
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import osmosis.folder.inspector.constants.Constant;
-import osmosis.folder.inspector.constants.ResourcePaths;
-
-import java.util.Objects;
-
-public class Main extends Application {
-    @Override
-    public void start(Stage primaryStage) throws Exception {
-        Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource(ResourcePaths.MAIN_FXML)));
-        primaryStage.setTitle(Constant.FOLDER_INSPECTOR);
-        primaryStage.setScene(new Scene(root));
-        primaryStage.show();
+public class Main {
+    private Main() {
     }
 
     public static void main(String[] args) {
-        launch(args);
+        FolderInspector.main(args);
     }
 }
